potpie/backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
httpx
python-dotenv
loguru
